requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
httpx==0.27.0
websockets==12.0
python-dotenv==1.0.1
prometheus-client==0.20.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
structlog==24.1.0
tenacity==8.3.0
sqlalchemy==2.0.31
ccxt==4.3.89
numpy==1.26.4